+++ app/src/modules/settings/index.ts
@@ -13,9 +13,9 @@ import Extensions from './routes/extensions/extensions.vue';
 import FlowOperationDetail from './routes/flows/components/operation-detail.vue';
 import FlowsDetail from './routes/flows/flow.vue';
 import FlowsOverview from './routes/flows/overview.vue';
-import MarketplaceExtension from './routes/marketplace/extension/extension.vue';
-import MarketplaceRegistry from './routes/marketplace/registry/registry.vue';
-import MarketplaceAccount from './routes/marketplace/account/account.vue';
+import MarketplaceAccount from './routes/marketplace/routes/account/account.vue';
+import MarketplaceExtension from './routes/marketplace/routes/extension/extension.vue';
+import MarketplaceRegistry from './routes/marketplace/routes/registry/registry.vue';
 import NotFound from './routes/not-found.vue';
 import PresetsCollection from './routes/presets/collection/collection.vue';
 import PresetsItem from './routes/presets/item.vue';
