+++ app/src/modules/settings/routes/marketplace/components/metadata-item.vue
