--- app/src/modules/settings/routes/marketplace/extension/components/extension-metadata-version.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import ExtensionMetadataItem from './extension-metadata-item.vue';
-
-defineProps<{ version: string }>();
-</script>
-
-<template>
-	<ExtensionMetadataItem icon="info">v{{ version }}</ExtensionMetadataItem>
-</template>
