+++ app/src/modules/settings/routes/marketplace/routes/extension/components/extension-banner.vue
