+++ app/src/modules/settings/routes/marketplace/routes/extension/components/extension-metadata-date.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { localizedFormatDistanceStrict } from '@/utils/localized-format-distance-strict';
 import { useI18n } from 'vue-i18n';
-import ExtensionMetadataItem from './extension-metadata-item.vue';
+import MetadataItem from '../../../components/metadata-item.vue';
 
 defineProps<{ publishDate: string }>();
 
@@ -9,7 +9,7 @@ const { t } = useI18n();
 </script>
 
 <template>
-	<ExtensionMetadataItem icon="event">
+	<MetadataItem icon="event">
 		{{
 			t('last_updated_relative', {
 				relativeTime: localizedFormatDistanceStrict(new Date(publishDate), new Date(), {
@@ -17,5 +17,5 @@ const { t } = useI18n();
 				}),
 			})
 		}}
-	</ExtensionMetadataItem>
+	</MetadataItem>
 </template>
