+++ app/src/modules/settings/routes/marketplace/routes/extension/components/extension-metadata-size.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { formatFilesize } from '@/utils/format-filesize';
 import { useI18n } from 'vue-i18n';
-import ExtensionMetadataItem from './extension-metadata-item.vue';
+import MetadataItem from '../../../components/metadata-item.vue';
 
 const { t } = useI18n();
 
@@ -9,7 +9,7 @@ defineProps<{ fileCount: number; unpackedSize: number }>();
 </script>
 
 <template>
-	<ExtensionMetadataItem icon="folder_open">
+	<MetadataItem icon="folder_open">
 		{{ formatFilesize(unpackedSize) }} ({{ t('n_files', fileCount) }})
-	</ExtensionMetadataItem>
+	</MetadataItem>
 </template>
