+++ app/src/modules/settings/routes/marketplace/routes/extension/components/extension-metadata-version.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import MetadataItem from '../../../components/metadata-item.vue';
+
+defineProps<{ version: string }>();
+</script>
+
+<template>
+	<MetadataItem icon="info">v{{ version }}</MetadataItem>
+</template>
