+++ app/src/modules/settings/routes/marketplace/routes/extension/components/extension-metadata.vue
@@ -6,7 +6,7 @@ import ExtensionMetadataAuthor from './extension-metadata-author.vue';
 import ExtensionMetadataCompatibility from './extension-metadata-compatibility.vue';
 import ExtensionMetadataDate from './extension-metadata-date.vue';
 import ExtensionMetadataDownloads from './extension-metadata-downloads.vue';
-import ExtensionMetadataItem from './extension-metadata-item.vue';
+import MetadataItem from '../../../components/metadata-item.vue';
 import ExtensionMetadataSize from './extension-metadata-size.vue';
 import ExtensionMetadataVersion from './extension-metadata-version.vue';
 
@@ -29,21 +29,21 @@ const latestVersion = computed(() => props.extension.versions.at(0)!);
 					:verified="latestVersion.publisher.verified"
 					:username="latestVersion.publisher.username"
 					:github-name="latestVersion.publisher.github_name"
-               :github-avatar-url="latestVersion.publisher.github_avatar_url"
+					:github-avatar-url="latestVersion.publisher.github_avatar_url"
 				/>
 				<ExtensionMetadataVersion :version="latestVersion.version" />
 				<ExtensionMetadataDownloads :downloads="extension.downloads" />
 				<ExtensionMetadataDate :publish-date="latestVersion.publish_date" />
 				<ExtensionMetadataSize :unpacked-size="latestVersion.unpacked_size" :file-count="latestVersion.file_count" />
-				<ExtensionMetadataItem v-if="latestVersion.url_homepage" icon="link" :href="latestVersion.url_homepage">
+				<MetadataItem v-if="latestVersion.url_homepage" icon="link" :href="latestVersion.url_homepage">
 					{{ t('homepage') }}
-				</ExtensionMetadataItem>
-				<ExtensionMetadataItem v-if="latestVersion.url_repository" icon="commit" :href="latestVersion.url_repository">
+				</MetadataItem>
+				<MetadataItem v-if="latestVersion.url_repository" icon="commit" :href="latestVersion.url_repository">
 					{{ t('repository') }}
-				</ExtensionMetadataItem>
-				<ExtensionMetadataItem v-if="latestVersion.url_bugs" icon="bug_report" :href="latestVersion.url_bugs">
+				</MetadataItem>
+				<MetadataItem v-if="latestVersion.url_bugs" icon="bug_report" :href="latestVersion.url_bugs">
 					{{ t('report_an_issue') }}
-				</ExtensionMetadataItem>
+				</MetadataItem>
 			</div>
 		</v-list>
 	</div>
