+++ app/src/modules/settings/routes/marketplace/routes/extension/extension.vue
@@ -4,10 +4,13 @@ import type { RegistryDescribeResponse } from '@directus/extensions-registry';
 import { ref, watchEffect } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
-import SettingsNavigation from '../../../components/navigation.vue';
+import SettingsNavigation from '../../../../components/navigation.vue';
+import ExtensionBanner from './components/extension-banner.vue';
+import ExtensionMetadata from './components/extension-metadata.vue';
+import ExtensionReadme from './components/extension-readme.vue';
 
 const props = defineProps<{
-	accountId: string;
+	extensionId: string;
 }>();
 
 const router = useRouter();
@@ -15,16 +18,16 @@ const { t } = useI18n();
 
 const loading = ref(false);
 const error = ref<unknown>(null);
-const account = ref<RegistryDescribeResponse>();
+const extension = ref<RegistryDescribeResponse['data']>();
 
 watchEffect(async () => {
-	if (!props.accountId) return;
+	if (!props.extensionId) return;
 
 	loading.value = true;
 
 	try {
-		const response = await api.get(`/extensions/registry/account/${props.accountId}`);
-		account.value = response.data.data;
+		const response = await api.get(`/extensions/registry/extension/${props.extensionId}`);
+		extension.value = response.data.data;
 	} catch (err) {
 		error.value = err;
 	} finally {
@@ -32,16 +35,7 @@ watchEffect(async () => {
 	}
 });
 
-const navigateBack = () => {
-	const backState = router.options.history.state.back;
-
-	if (typeof backState !== 'string' || !backState.startsWith('/login')) {
-		router.back();
-		return;
-	}
-
-	router.push('/settings/marketplace');
-};
+const navigateBack = () => router.push('/settings/marketplace');
 </script>
 
 <template>
@@ -56,10 +50,14 @@ const navigateBack = () => {
 			<settings-navigation />
 		</template>
 
-		<div class="account-content">
-			<template v-if="account">
+		<div class="extension-content">
+			<template v-if="extension">
 				<div class="container">
-					<div class="grid"></div>
+					<div class="grid">
+						<ExtensionBanner class="banner" :extension="extension" />
+						<ExtensionMetadata class="metadata" :extension="extension" />
+						<ExtensionReadme v-if="extension.readme" class="readme" :readme="extension.readme" />
+					</div>
 				</div>
 			</template>
 
@@ -71,7 +69,7 @@ const navigateBack = () => {
 </template>
 
 <style scoped lang="scss">
-.account-content {
+.extension-content {
 	padding: var(--content-padding);
 	padding-bottom: var(--content-padding-bottom);
 	max-width: 1200px;
@@ -109,4 +107,3 @@ const navigateBack = () => {
 	}
 }
 </style>
-
