+++ app/src/modules/settings/routes/marketplace/routes/registry/components/extension-list-item.vue
