+++ app/src/modules/settings/routes/marketplace/routes/registry/components/type-filter.vue
